pyproject.toml
[project]
name = "vibe-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.67.0",
    "ruff>=0.15.6",
    "streamlit>=1.55.0",
]
